--- lib/controller.dart
@@ -1,217 +1,0 @@
-import 'dart:async';
-import 'dart:io';
-
-import 'package:fl_clash/common/common.dart';
-import 'package:fl_clash/enum/enum.dart';
-import 'package:fl_clash/providers/providers.dart';
-import 'package:fl_clash/state.dart';
-import 'package:fl_clash/widgets/dialog.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:url_launcher/url_launcher.dart';
-
-class AppController {
-  late final BuildContext _context;
-  late final WidgetRef _ref;
-  bool isAttach = false;
-
-  static AppController? _instance;
-
-  AppController._internal();
-
-  factory AppController() {
-    _instance ??= AppController._internal();
-    return _instance!;
-  }
-
-  Future<void> attach(BuildContext context, WidgetRef ref) async {
-    _context = context;
-    _ref = ref;
-    await _init();
-    isAttach = true;
-  }
-
-  Future<void> _init() async {
-    FlutterError.onError = (details) {
-      commonPrint.log(
-        'exception: ${details.exception} stack: ${details.stack}',
-        logLevel: LogLevel.warning,
-      );
-    };
-    _ref.read(systemActionProvider.notifier).updateTray();
-    _ref.read(profilesActionProvider.notifier).autoUpdateProfiles();
-    _autoCheckUpdate();
-    autoLaunch?.updateStatus(_ref.read(appSettingProvider).autoLaunch);
-    if (!_ref.read(appSettingProvider).silentLaunch) {
-      window?.show();
-    } else {
-      window?.hide();
-    }
-    await _handleFailedPreference();
-    await _handlerDisclaimer();
-    await _showCrashlyticsTip();
-    await _ref.read(coreActionProvider.notifier).connectCore();
-    await _ref.read(coreActionProvider.notifier).initCore();
-    await _initStatus();
-    _ref.read(initProvider.notifier).value = true;
-  }
-
-  Future<void> _handleFailedPreference() async {
-    if (await preferences.isInit) return;
-    final res = await globalState.showMessage(
-      title: currentAppLocalizations.tip,
-      message: TextSpan(text: currentAppLocalizations.cacheCorrupt),
-    );
-    if (res == true) {
-      final file = File(await appPath.sharedPreferencesPath);
-      await file.safeDelete();
-    }
-    await _ref.read(systemActionProvider.notifier).handleExit();
-  }
-
-  void initLink() {
-    linkManager.initAppLinksListen((url) async {
-      final res = await globalState.showMessage(
-        title: currentAppLocalizations.addProfile,
-        message: TextSpan(
-          children: [
-            TextSpan(text: currentAppLocalizations.doYouWantToPass),
-            TextSpan(
-              text: ' $url ',
-              style: TextStyle(
-                color: _context.colorScheme.primary,
-                decoration: TextDecoration.underline,
-                decorationColor: _context.colorScheme.primary,
-              ),
-            ),
-            TextSpan(text: currentAppLocalizations.createProfile),
-          ],
-        ),
-      );
-      if (res != true) return;
-      _ref.read(profilesActionProvider.notifier).addProfileFormURL(url);
-    });
-  }
-
-  Future<bool> showDisclaimer() async {
-    return await globalState.showCommonDialog<bool>(
-          dismissible: false,
-          child: CommonDialog(
-            title: currentAppLocalizations.disclaimer,
-            actions: [
-              TextButton(
-                onPressed: () {
-                  Navigator.of(_context).pop<bool>(false);
-                },
-                child: Text(currentAppLocalizations.exit),
-              ),
-              TextButton(
-                onPressed: () {
-                  Navigator.of(_context).pop<bool>(true);
-                },
-                child: Text(currentAppLocalizations.agree),
-              ),
-            ],
-            child: Text(currentAppLocalizations.disclaimerDesc),
-          ),
-        ) ??
-        false;
-  }
-
-  Future<void> _showCrashlyticsTip() async {
-    if (!system.isAndroid) return;
-    if (_ref.read(appSettingProvider.select((state) => state.crashlyticsTip))) {
-      return;
-    }
-    await globalState.showMessage(
-      title: currentAppLocalizations.dataCollectionTip,
-      cancelable: false,
-      message: TextSpan(text: currentAppLocalizations.dataCollectionContent),
-    );
-    _ref
-        .read(appSettingProvider.notifier)
-        .update((state) => state.copyWith(crashlyticsTip: true));
-  }
-
-  Future<void> _handlerDisclaimer() async {
-    if (_ref.read(
-      appSettingProvider.select((state) => state.disclaimerAccepted),
-    )) {
-      return;
-    }
-    final isDisclaimerAccepted = await showDisclaimer();
-    if (!isDisclaimerAccepted) {
-      await _ref.read(systemActionProvider.notifier).handleExit();
-    }
-    _ref
-        .read(appSettingProvider.notifier)
-        .update((state) => state.copyWith(disclaimerAccepted: true));
-  }
-
-  Future<void> _initStatus() async {
-    if (!globalState.needInitStatus) {
-      commonPrint.log('init status cancel');
-      return;
-    }
-    commonPrint.log('init status');
-    if (system.isAndroid) {
-      await globalState.updateStartTime();
-    }
-    final status = globalState.isStart == true
-        ? true
-        : _ref.read(appSettingProvider).autoRun;
-    if (status == true) {
-      await _ref
-          .read(setupActionProvider.notifier)
-          .updateStatus(true, isInit: true);
-    } else {
-      await _ref.read(setupActionProvider.notifier).applyProfile(force: true);
-    }
-  }
-
-  Future<void> _autoCheckUpdate() async {
-    if (!_ref.read(appSettingProvider).autoCheckUpdate) return;
-    final res = await request.checkForUpdate();
-    checkUpdateResultHandle(data: res);
-  }
-
-  Future<void> checkUpdateResultHandle({
-    Map<String, dynamic>? data,
-    bool isUser = false,
-  }) async {
-    if (data != null) {
-      final tagName = data['tag_name'];
-      final body = data['body'];
-      final submits = utils.parseReleaseBody(body);
-      final textTheme = _context.textTheme;
-      final res = await globalState.showMessage(
-        title: currentAppLocalizations.discoverNewVersion,
-        message: TextSpan(
-          text: '$tagName \n',
-          style: textTheme.headlineSmall,
-          children: [
-            TextSpan(text: '\n', style: textTheme.bodyMedium),
-            for (final submit in submits)
-              TextSpan(text: '- $submit \n', style: textTheme.bodyMedium),
-          ],
-        ),
-        confirmText: currentAppLocalizations.goDownload,
-        cancelText: isUser ? null : currentAppLocalizations.noLongerRemind,
-      );
-      if (res == true) {
-        launchUrl(Uri.parse('https://github.com/$repository/releases/latest'));
-      } else if (!isUser && res == false) {
-        _ref
-            .read(appSettingProvider.notifier)
-            .update((state) => state.copyWith(autoCheckUpdate: false));
-      }
-    } else if (isUser) {
-      globalState.showMessage(
-        title: currentAppLocalizations.checkUpdate,
-        message: TextSpan(text: currentAppLocalizations.checkUpdateError),
-      );
-    }
-  }
-}
-
-final appController = AppController();
