+++ lib/state.dart
@@ -1,15 +1,13 @@
 import 'dart:async';
 import 'dart:convert';
 import 'dart:ffi' as ffi;
+import 'dart:io';
 
 import 'package:animations/animations.dart';
 import 'package:dynamic_color/dynamic_color.dart';
 import 'package:fl_clash/common/theme.dart';
 import 'package:fl_clash/core/core.dart';
 import 'package:fl_clash/plugins/service.dart';
-import 'package:fl_clash/providers/app.dart';
-import 'package:fl_clash/providers/config.dart';
-import 'package:fl_clash/providers/database.dart';
 import 'package:fl_clash/widgets/dialog.dart';
 import 'package:fl_clash/widgets/list.dart';
 import 'package:flutter/material.dart';
@@ -26,6 +24,7 @@ import 'database/database.dart';
 import 'enum/enum.dart';
 import 'l10n/l10n.dart';
 import 'models/models.dart';
+import 'providers/providers.dart';
 
 typedef UpdateTasks = List<FutureOr Function()>;
 
@@ -49,6 +48,7 @@ class GlobalState {
   UpdateTasks tasks = [];
   String? lastConfigMd5;
   VpnState? lastVpnState;
+  bool isAttach = false;
 
   bool get isStart => startTime != null && startTime!.isBeforeNow;
 
@@ -78,6 +78,8 @@ class GlobalState {
   String get ua => container
       .read(patchClashConfigProvider.select((state) => state.globalUa))
       .takeFirstValid([packageInfo.ua]);
+
+  BuildContext get _context => navigatorKey.currentContext!;
 
   Future<ProviderContainer> _initData(int version) async {
     final appState = AppState(
@@ -345,6 +347,197 @@ class GlobalState {
       false => Map<String, dynamic>.from(res.rawResult),
     };
     return value ?? config;
+  }
+
+  Future<void> attach(BuildContext context, WidgetRef ref) async {
+    await _initApp();
+    isAttach = true;
+  }
+
+  Future<void> _initApp() async {
+    FlutterError.onError = (details) {
+      commonPrint.log(
+        'exception: ${details.exception} stack: ${details.stack}',
+        logLevel: LogLevel.warning,
+      );
+    };
+    container.read(systemActionProvider.notifier).updateTray();
+    container.read(profilesActionProvider.notifier).autoUpdateProfiles();
+    _autoCheckUpdate();
+    autoLaunch?.updateStatus(container.read(appSettingProvider).autoLaunch);
+    if (!container.read(appSettingProvider).silentLaunch) {
+      window?.show();
+    } else {
+      window?.hide();
+    }
+    await _handleFailedPreference();
+    await _handlerDisclaimer();
+    await _showCrashlyticsTip();
+    await container.read(coreActionProvider.notifier).connectCore();
+    await container.read(coreActionProvider.notifier).initCore();
+    await _initStatus();
+    container.read(initProvider.notifier).value = true;
+  }
+
+  Future<void> _handleFailedPreference() async {
+    if (await preferences.isInit) return;
+    final res = await showMessage(
+      title: currentAppLocalizations.tip,
+      message: TextSpan(text: currentAppLocalizations.cacheCorrupt),
+    );
+    if (res == true) {
+      final file = File(await appPath.sharedPreferencesPath);
+      await file.safeDelete();
+    }
+    await container.read(systemActionProvider.notifier).handleExit();
+  }
+
+  void initLink() {
+    linkManager.initAppLinksListen((url) async {
+      final res = await showMessage(
+        title: currentAppLocalizations.addProfile,
+        message: TextSpan(
+          children: [
+            TextSpan(text: currentAppLocalizations.doYouWantToPass),
+            TextSpan(
+              text: ' $url ',
+              style: TextStyle(
+                color: _context.colorScheme.primary,
+                decoration: TextDecoration.underline,
+                decorationColor: _context.colorScheme.primary,
+              ),
+            ),
+            TextSpan(text: currentAppLocalizations.createProfile),
+          ],
+        ),
+      );
+      if (res != true) return;
+      container.read(profilesActionProvider.notifier).addProfileFormURL(url);
+    });
+  }
+
+  Future<bool> showDisclaimer() async {
+    return await showCommonDialog<bool>(
+          dismissible: false,
+          child: CommonDialog(
+            title: currentAppLocalizations.disclaimer,
+            actions: [
+              TextButton(
+                onPressed: () {
+                  Navigator.of(_context).pop<bool>(false);
+                },
+                child: Text(currentAppLocalizations.exit),
+              ),
+              TextButton(
+                onPressed: () {
+                  Navigator.of(_context).pop<bool>(true);
+                },
+                child: Text(currentAppLocalizations.agree),
+              ),
+            ],
+            child: Text(currentAppLocalizations.disclaimerDesc),
+          ),
+        ) ??
+        false;
+  }
+
+  Future<void> _showCrashlyticsTip() async {
+    if (!system.isAndroid) return;
+    if (container.read(
+      appSettingProvider.select((state) => state.crashlyticsTip),
+    )) {
+      return;
+    }
+    await showMessage(
+      title: currentAppLocalizations.dataCollectionTip,
+      cancelable: false,
+      message: TextSpan(text: currentAppLocalizations.dataCollectionContent),
+    );
+    container
+        .read(appSettingProvider.notifier)
+        .update((state) => state.copyWith(crashlyticsTip: true));
+  }
+
+  Future<void> _handlerDisclaimer() async {
+    if (container.read(
+      appSettingProvider.select((state) => state.disclaimerAccepted),
+    )) {
+      return;
+    }
+    final isDisclaimerAccepted = await showDisclaimer();
+    if (!isDisclaimerAccepted) {
+      await container.read(systemActionProvider.notifier).handleExit();
+    }
+    container
+        .read(appSettingProvider.notifier)
+        .update((state) => state.copyWith(disclaimerAccepted: true));
+  }
+
+  Future<void> _initStatus() async {
+    if (!needInitStatus) {
+      commonPrint.log('init status cancel');
+      return;
+    }
+    commonPrint.log('init status');
+    if (system.isAndroid) {
+      await updateStartTime();
+    }
+    final status = isStart == true
+        ? true
+        : container.read(appSettingProvider).autoRun;
+    if (status == true) {
+      await container
+          .read(setupActionProvider.notifier)
+          .updateStatus(true, isInit: true);
+    } else {
+      await container
+          .read(setupActionProvider.notifier)
+          .applyProfile(force: true);
+    }
+  }
+
+  Future<void> _autoCheckUpdate() async {
+    if (!container.read(appSettingProvider).autoCheckUpdate) return;
+    final res = await request.checkForUpdate();
+    checkUpdateResultHandle(data: res);
+  }
+
+  Future<void> checkUpdateResultHandle({
+    Map<String, dynamic>? data,
+    bool isUser = false,
+  }) async {
+    if (data != null) {
+      final tagName = data['tag_name'];
+      final body = data['body'];
+      final submits = utils.parseReleaseBody(body);
+      final textTheme = _context.textTheme;
+      final res = await showMessage(
+        title: currentAppLocalizations.discoverNewVersion,
+        message: TextSpan(
+          text: '$tagName \n',
+          style: textTheme.headlineSmall,
+          children: [
+            TextSpan(text: '\n', style: textTheme.bodyMedium),
+            for (final submit in submits)
+              TextSpan(text: '- $submit \n', style: textTheme.bodyMedium),
+          ],
+        ),
+        confirmText: currentAppLocalizations.goDownload,
+        cancelText: isUser ? null : currentAppLocalizations.noLongerRemind,
+      );
+      if (res == true) {
+        launchUrl(Uri.parse('https://github.com/$repository/releases/latest'));
+      } else if (!isUser && res == false) {
+        container
+            .read(appSettingProvider.notifier)
+            .update((state) => state.copyWith(autoCheckUpdate: false));
+      }
+    } else if (isUser) {
+      showMessage(
+        title: currentAppLocalizations.checkUpdate,
+        message: TextSpan(text: currentAppLocalizations.checkUpdateError),
+      );
+    }
   }
 }
