+++ lib/views/about.dart
@@ -1,7 +1,6 @@
 import 'dart:async';
 
 import 'package:fl_clash/common/common.dart';
-import 'package:fl_clash/controller.dart';
 import 'package:fl_clash/l10n/l10n.dart';
 import 'package:fl_clash/providers/config.dart';
 import 'package:fl_clash/state.dart';
@@ -31,7 +30,7 @@ class AboutView extends StatelessWidget {
       request.checkForUpdate,
       title: context.appLocalizations.checkUpdate,
     );
-    appController.checkUpdateResultHandle(data: data, isUser: true);
+    globalState.checkUpdateResultHandle(data: data, isUser: true);
   }
 
   List<Widget> _buildMoreSection(BuildContext context) {
